Refactor: fcm 관련 함수 재배치
firebase.ts에 넣음
